+++ scripts/actions/alchemical-crossbow.js
@@ -1,7 +1,7 @@
 import { HookManager } from "../utils/hook-manager.js";
 import { dialogPrompt } from "../utils/prompt-dialog.js";
 import { Updates } from "../utils/updates.js";
-import { findItemOnActor, getControlledActorAndToken, getEffectFromActor, getItem, postToChat, setChoice, setEffectTarget, showWarning } from "../utils/utils.js";
+import { findItemOnActor, getControlledActorAndToken, getEffectFromActor, getItem, postInteractToChat, setChoice, setEffectTarget, showWarning } from "../utils/utils.js";
 import { getWeapon } from "../utils/weapon-utils.js";
 
 const LOADED_BOMB_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.Item.cA9sBCFAxY2EJgrC";
@@ -88,11 +88,10 @@ export async function loadAlchemicalCrossbow() {
     // Remove one bomb from the stack
     updates.update(bomb, { "system.quantity": bomb.quantity - 1 });
 
-    await postToChat(
+    await postInteractToChat(
         actor,
         bomb.img,
         format("tokenLoadsWeaponWithBomb", { token: token.name, weapon: weapon.name, bomb: bomb.name }),
-        game.i18n.localize("PF2E.Actions.Interact.Title"),
         "1"
     );
 
@@ -140,7 +139,7 @@ export async function unloadAlchemicalCrossbow() {
     await unloadBomb(actor, loadedBombEffect, updates);
 
     if (loadedBombFlags.bombCharges > 0) {
-        await postToChat(
+        await postInteractToChat(
             actor,
             UNLOAD_BOMB_IMG,
             format(
@@ -151,7 +150,6 @@ export async function unloadAlchemicalCrossbow() {
                     bomb: loadedBombFlags.bombName
                 }
             ),
-            game.i18n.localize("PF2E.Actions.Interact.Title"),
             "1"
         );
     }