+++ scripts/ammunition-system/actions/next-chamber.js
@@ -2,7 +2,7 @@ import { Weapon } from "../../types/pf2e-ranged-combat/weapon.js";
 import { PF2eActor } from "../../types/pf2e/actor.js";
 import { PF2eToken } from "../../types/pf2e/token.js";
 import { Updates } from "../../utils/updates.js";
-import { getControlledActorAndToken, getEffectFromActor, getFlag, getItem, postToChat, setEffectTarget, showWarning, useAdvancedAmmunitionSystem } from "../../utils/utils.js";
+import { getControlledActorAndToken, getEffectFromActor, getFlag, getItem, postInteractToChat, setEffectTarget, showWarning, useAdvancedAmmunitionSystem } from "../../utils/utils.js";
 import { getWeapon } from "../../utils/weapon-utils.js";
 import { CHAMBER_LOADED_EFFECT_ID, CONJURED_ROUND_ITEM_ID, SELECT_NEXT_CHAMBER_IMG } from "../constants.js";
 import { getSelectedAmmunition, isLoaded } from "../utils.js";
@@ -58,11 +58,10 @@ export async function performNextChamber(actor, token, weapon) {
         }
 
         await setLoadedChamber(actor, weapon, selectedAmmunition, updates);
-        await postToChat(
+        await postInteractToChat(
             token.actor,
             SELECT_NEXT_CHAMBER_IMG,
             format("chatMessageSelectChamber", { token: token.name, ammunition: selectedAmmunition.name, weapon: weapon.name }),
-            game.i18n.localize("PF2E.Actions.Interact.Title"),
             1,
         );
     } else {
@@ -73,11 +72,10 @@ export async function performNextChamber(actor, token, weapon) {
         }
 
         await addChamberLoaded(actor, weapon, null, updates);
-        await postToChat(
+        await postInteractToChat(
             token.actor,
             SELECT_NEXT_CHAMBER_IMG,
             format("chatMessageSelectNextChamber", { token: token.name, weapon: weapon.name }),
-            game.i18n.localize("PF2E.Actions.Interact.Title"),
             1,
         );
     }