+++ scripts/ammunition-system/auxiliary-actions.js
@@ -345,7 +345,7 @@ async function changeCarryType(weapon, subAction, hands) {
                 subtitle: `PF2E.Actions.Interact.${subAction}.Title`,
                 glyph: "1",
             },
-            traits: [traitSlugToObject("manipulate")]
+            traits: ["manipulate"]
         }
     );
     const content = await renderTemplate(